Fix v5 protocol encoding bugs, QoS 2 race, and zero-copy PUBLISH path
- Fix FOUR_BYTE_INTEGER properties (0x11 SessionExpiryInterval, 0x27
  MaximumPacketSize) writing 8 bytes instead of 4: use toUInt() in
  size() and write() methods, fix reads to use readUnsignedInt()
- Fix v5 ACK packets (PUBACK, PUBREC, PUBREL, PUBCOMP) failing to
  parse when remainingLength == 3 (reason code without properties)
- Fix v5 DISCONNECT failing to parse when remainingLength is 0 (no
  fields) or 1 (reason code only) — pass remainingLength through
- Fix zero-copy PUBLISH missing v5 properties in serializeHeaderToSlice:
  serialize v5 properties after topic + packetId
- Fix MqttSocketSession header buffer sizing: use packet.packetSize()
  minus payload instead of hardcoded 64 bytes
- Fix QoS 2 observer race: split publish into preparePublish → observePub
  → sendPacket so SharedFlow collectors start before packet hits wire
- Remove pubRecReceived.await() gate from PUBCOMP observer
- Switch integration tests from runTest to runTestNoTimeSkipping for
  real network I/O (virtual time breaks socket operations)
- Increase MaxDirectMemorySize to 2g for TLS integration tests
- Update v5 unit test assertions for 4-byte property encoding

+++ models-v5/src/commonMain/kotlin/com/ditchoom/mqtt5/controlpacket/properties/MaximumPacketSize.kt
@@ -5,7 +5,7 @@ import com.ditchoom.buffer.WriteBuffer
 data class MaximumPacketSize(
     val packetSizeLimitationBytes: ULong,
 ) : Property(0x27, Type.FOUR_BYTE_INTEGER) {
-    override fun size(): Int = size(packetSizeLimitationBytes)
+    override fun size(): Int = size(packetSizeLimitationBytes.toUInt())
 
-    override fun write(buffer: WriteBuffer): Int = write(buffer, packetSizeLimitationBytes)
+    override fun write(buffer: WriteBuffer): Int = write(buffer, packetSizeLimitationBytes.toUInt())
 }
